pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "lean_reinforcement"
version = "0.1.0"
description = "Reinforcement learning for Lean theorem proving"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "lean-dojo",
    "wandb",
    "pytorch-lightning",
    "transformers",
    "gymnasium",
    "loguru",
    "networkx",
    "lxml",
    "ReProver @ file://./ReProver"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["lean_reinforcement*"]
